[refactoring] EraseCommand 에 range 범위 체크부분을 추가하였습니다
