+++ TestConsole/src/main/java/command/common/Command.java
@@ -3,21 +3,14 @@ package command.common;
 import logger.Logger;
 
 public abstract class Command implements ICommand {
+
     public static final String ERROR_FLAG = "ERROR";
     public static final String PASS_FLAG = "PASS";
     public static final String FAIL_FLAG = "FAIL";
+
+    protected CommandValidator validator;
     protected Logger logger = Logger.getInstance();
     protected final ConsoleService service;
-
-    public static final String NO_NEED_TO_VALID_CHECK = "";
-
-    public static final String VALID_ARGUMENT = "";
-    public static final String VALID_ADDRESS = "";
-    public static final String VALID_DATA_FORMAT = "";
-    public static final String INVALID_ARGUMENT_NUMBER_MSG = "ERROR : Invalid Argument Number !";
-    public static final String INVALID_ADDRESS_FORMAT_MSG = "ERROR : LBA must be between 0 and 99.";
-    public static final String INVALID_DATA_FORMAT = "ERROR Value must be in hex format (e.g., 0x1234ABCD)";
-    public static final String DATA_FORMAT = "^0x[0-9A-Fa-f]{8}$";
 
     public Command(ConsoleService service) {
         this.service =  service ;
@@ -32,31 +25,10 @@ public abstract class Command implements ICommand {
         return doExecute(args);
     }
 
-    abstract public String argumentsValidCheck(String[] args);
+    public String argumentsValidCheck(String[] args){
+        return validator.validCheck(args);
+    }
     abstract public CommandResult doExecute(String[] args);
-
-
-    public boolean isInValidArgumentCount(String[] args, int expected) {
-        return args.length != expected;
-    }
-
-    public String addressValidCheck(String address) {
-        try {
-            int lba = Integer.parseInt(address);
-            if (lba < 0 || lba > 99) {
-                return INVALID_ADDRESS_FORMAT_MSG;
-            }
-        }
-        catch (NumberFormatException e) {
-            return "ERROR : NumberFormatException" + e.getMessage();
-        }
-
-        return VALID_ADDRESS;
-    }
-
-    public boolean isInValidData(String data) {
-        return !data.matches(DATA_FORMAT);
-    }
 
     public boolean isErrorExist(String result){
         return result.startsWith(ERROR_FLAG);
