+++ TestConsole/src/main/java/command/common/ICommandValidator.java
@@ -1,0 +1,5 @@
+package command.common;
+
+public interface ICommandValidator {
+    String validCheck(String[] args);
+}
