+++ TestConsole/src/main/java/command/console/actor/ExitCommand.java
@@ -1,19 +1,14 @@
-package command.console;
+package command.console.actor;
 
 import command.common.Command;
 import command.common.CommandResult;
 import command.common.ConsoleService;
+import command.console.validator.ExitCommandValidator;
 
 public class ExitCommand extends Command {
     public ExitCommand(ConsoleService service) {
         super(service);
-    }
-    public static final int EXPECTED_ARGUMENT_COUNT = 1;
-
-    @Override
-    public String argumentsValidCheck(String[] args) {
-        if(isInValidArgumentCount(args, EXPECTED_ARGUMENT_COUNT)) return INVALID_ARGUMENT_NUMBER_MSG;
-        return VALID_ARGUMENT;
+        validator = new ExitCommandValidator();
     }
 
     @Override
