+++ TestConsole/src/main/java/command/console/actor/FlushCommand.java
@@ -1,21 +1,14 @@
-package command.console;
+package command.console.actor;
 
 import command.common.Command;
 import command.common.CommandResult;
 import command.common.ConsoleService;
+import command.console.validator.FlushCommandValid;
 
 public class FlushCommand extends Command {
-
-    public static final int EXPECTED_ARGUMENT_COUNT = 1;
-
     public FlushCommand(ConsoleService service) {
         super(service);
-    }
-
-    @Override
-    public String argumentsValidCheck(String[] args) {
-        if(isInValidArgumentCount(args, EXPECTED_ARGUMENT_COUNT)) return INVALID_ARGUMENT_NUMBER_MSG;
-        return VALID_ARGUMENT;
+        validator = new FlushCommandValid();
     }
 
     @Override
