+++ TestConsole/src/main/java/command/console/actor/FullReadCommand.java
@@ -1,20 +1,16 @@
-package command.console;
+package command.console.actor;
 
 import command.common.Command;
 import command.common.CommandResult;
 import command.common.ConsoleService;
+import command.console.validator.FullReadCommandValidator;
 
 public class FullReadCommand extends Command {
     public static final int EXPECTED_ARGUMENT_COUNT = 1;
 
     public FullReadCommand(ConsoleService service) {
         super(service);
-    }
-    
-    @Override
-    public String argumentsValidCheck(String[] args) {
-        if(isInValidArgumentCount(args, EXPECTED_ARGUMENT_COUNT)) return INVALID_ARGUMENT_NUMBER_MSG;
-        return VALID_ARGUMENT;
+        validator = new FullReadCommandValidator();
     }
 
     @Override
