+++ TestConsole/src/main/java/command/console/actor/FullWriteCommand.java
@@ -1,21 +1,14 @@
-package command.console;
+package command.console.actor;
 
 import command.common.Command;
 import command.common.CommandResult;
 import command.common.ConsoleService;
+import command.console.validator.FullWriteCommandValidator;
 
 public class FullWriteCommand extends Command {
-
-    public static final int EXPECTED_ARGUMENT_COUNT = 2;
-
     public FullWriteCommand(ConsoleService service) {
         super(service);
-    }
-
-    public String argumentsValidCheck(String[] args) {
-        if(isInValidArgumentCount(args, EXPECTED_ARGUMENT_COUNT)) return INVALID_ARGUMENT_NUMBER_MSG;
-        if(isInValidData(args[1])) return INVALID_DATA_FORMAT;
-        return VALID_ARGUMENT;
+        validator = new FullWriteCommandValidator();
     }
 
     @Override
