+++ TestConsole/src/main/java/command/console/actor/ReadCommand.java
@@ -1,22 +1,14 @@
-package command.console;
+package command.console.actor;
 
 import command.common.Command;
 import command.common.CommandResult;
 import command.common.ConsoleService;
+import command.console.validator.ReadCommandValidator;
 
 public class ReadCommand extends Command {
-
-    public static final int EXPECTED_ARGUMENT_COUNT = 2;
-
     public ReadCommand(ConsoleService service) {
         super(service);
-    }
-
-    @Override
-    public String argumentsValidCheck(String[] args) {
-        if(isInValidArgumentCount(args, EXPECTED_ARGUMENT_COUNT)) return INVALID_ARGUMENT_NUMBER_MSG;
-        if(!addressValidCheck(args[1]).equals(VALID_ADDRESS)) return INVALID_ADDRESS_FORMAT_MSG;
-        return VALID_ARGUMENT;
+        validator = new ReadCommandValidator();
     }
 
     @Override
