+++ TestConsole/src/main/java/command/script/PartialLBAWrite.java
@@ -5,18 +5,13 @@ import command.common.CommandResult;
 import command.common.ConsoleService;
 
 //2_PartialLBAWrite
-public class PartialLBAWrite extends Command {
+public class PartialLBAWrite extends TestScript {
     public static final int LOOP_COUNT = 30;
     public static final String TEST_VALUE = "0x12345678";
     public static final int[] LBA_TEST_SEQUENCE = new int[]{4, 0, 3, 1, 2};
 
     public PartialLBAWrite(ConsoleService service) {
         super(service);
-    }
-
-    @Override
-    public String argumentsValidCheck(String[] args) {
-        return NO_NEED_TO_VALID_CHECK;
     }
 
     @Override
