+++ TestConsole/src/main/java/command/script/TestScript.java
@@ -1,0 +1,11 @@
+package command.script;
+
+import command.common.Command;
+import command.common.ConsoleService;
+
+public abstract class TestScript extends Command {
+    public TestScript(ConsoleService service) {
+        super(service);
+        validator = new TestScriptValidator();
+    }
+}
